Calculate cost initialized

+++ database/migrations/2020_12_01_124649_create_configurations_table.php
@@ -15,7 +15,7 @@ class CreateConfigurationsTable extends Migration
     {
         Schema::create('configurations', function (Blueprint $table) {
             $table->id();
-            $table->float('current_ipc');
+            $table->decimal('current_ipc');
             $table->tinyInteger('iva');
             $table->timestamps();
         });
